Replaced experimental_rerun() with rerun()
streamlit app is giving an error when you hit submit on the interviewer page that experimental rerun doesn't exist. I have gotten that error in the past whenever ai generates the code. Seems this function is not available in the versions of streamlit we are running. Using rerun() instead fixes it.